css-text/line-breaking/css-text-line-break-ns.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>Line-breaking: Non-starter punctuation</title>
<link rel="author" title="Richard Ishida" href="mailto:ishida@w3.org">
<link rel="help" href="https://drafts.csswg.org/css-text-3/#line-breaking">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<meta name="assert" content="[Exploratory] The browser will not allow a character with the NS (Non-starter) linebreak property at the beginning of a line.">
<style type="text/css">
@font-face {
    font-family: 'mplus-1p-regular';
    src: url('../../../fonts/CSSTest/mplus-1p-regular.woff') format('woff');
    }
#wrapper { position: relative; }
.test { color: black; }
.test, .ref { font-size: 30px; font-family: mplus-1p-regular, sans-serif; width: 185px; padding: 0; border: 1px solid orange; line-height: 1em; }
.ref { color: lightgray; border-color: white; margin-bottom: 1em; }
</style>
<style>
.test { line-break: loose; }
</style>
<script>
var charlist = `17D6 KHMER SIGN CAMNUC PII KUUH
16FE1 NUSHU ITERATION MARK
A015 YI SYLLABLE WU
16FE0 TANGUT ITERATION MARK
203C DOUBLE EXCLAMATION MARK
203D INTERROBANG
2047 DOUBLE QUESTION MARK
2048 QUESTION EXCLAMATION MARK
2049 EXCLAMATION QUESTION MARK
3005 IDEOGRAPHIC ITERATION MARK
301C WAVE DASH
303B VERTICAL IDEOGRAPHIC ITERATION MARK
303C MASU MARK
309B KATAKANA-HIRAGANA VOICED SOUND MARK
309C KATAKANA-HIRAGANA SEMI-VOICED SOUND MARK
309D HIRAGANA ITERATION MARK
309E HIRAGANA VOICED ITERATION MARK
30A0 KATAKANA-HIRAGANA DOUBLE HYPHEN
30FB KATAKANA MIDDLE DOT
30FD KATAKANA ITERATION MARK
30FE KATAKANA VOICED ITERATION MARK
FE54 SMALL SEMICOLON
FE55 SMALL COLON
FF1A FULLWIDTH COLON
FF1B FULLWIDTH SEMICOLON
FF65 HALFWIDTH KATAKANA MIDDLE DOT
FF9E HALFWIDTH KATAKANA VOICED SOUND MARK
FF9F HALFWIDTH KATAKANA SEMI-VOICED SOUND MARK
1F679 HEAVY INTERROBANG ORNAMENT
1F67A SANS-SERIF INTERROBANG ORNAMENT
1F67B HEAVY SANS-SERIF INTERROBANG ORNAMENT`
</script>
</head>
<body>
<script>
var lines = charlist.split('\n')
var out = '<div id="log"></div>\n'

// create the HTML
out += '<h2>Test output</h2><p>Test passes if the NS character wraps to the second line with another character before it (as shown by the lighter reference text).</p>\n'
for (var i=0;i<lines.length;i++) {
    // get the data
    var firstSpace = lines[i].indexOf(' ')
    var hex = lines[i].substr(0,firstSpace)
    var name = lines[i].substr(firstSpace)
    // make a test
    out +=  '<div class="wrapper"><div>'+hex+' '+name+'</div>' +
    '<div class="test" id="test'+i+'" lang="ja">文文文文文文&#x'+hex+';字<span id="testSpan'+i+'">字</span></div>' +
     '<div class="ref" id="ref'+i+'" lang="ja">文文文文文<br/>文&#x'+hex+';字<span id="refSpan'+i+'">字</span></div>' +
     '</div>'
    }
// create a section to list failures
out += '<h2>List of failures</h2><p id="failures">none</p>'
document.querySelector('body').innerHTML = out

// make a list of failures
var summary = ''
for (i=0;i<lines.length;i++) {
    if (document.getElementById('testSpan'+i).offsetLeft !== document.getElementById('refSpan'+i).offsetLeft) {
        if (summary !== '') summary += ', '
        summary += 'U+'+document.getElementById('test'+i).parentNode.firstChild.textContent
        }
    }

// run the test framework
 for (i=0;i<lines.length;i++) {
   test(function() {
        assert_true(document.getElementById('testSpan'+i).offsetLeft === document.getElementById('refSpan'+i).offsetLeft);
        }, lines[i]+' should not appear at the start of a line');
   }

// list unsuccessful tests
if (summary != '') document.getElementById('failures').textContent = summary
</script>
<!--Notes:
This list is based on property assignments for Unicode version 10.
The test creates a box width that causes the 8th character to wrap. It also creates the expected behaviour with a ref instance, using <br/>. Each line ends with a span. The test then checks whether the left edge of the span is in the same place in test and ref instance.
-->
</body>
</html>
